+++ x/common/error.go
@@ -187,3 +187,11 @@ func CombineErrorsFromStrings(strs ...string) (err error) {
 func ErrNilMsg() error {
 	return grpcstatus.Errorf(grpccodes.InvalidArgument, "nil msg")
 }
+
+// ErrNotImplemented: Represents an function error value.
+func ErrNotImplemented() error { return fmt.Errorf("fn not implemented yet") }
+
+// ErrNotImplementedGprc: Represents an unimplemented gRPC method.
+func ErrNotImplementedGprc() error {
+	return grpcstatus.Error(grpccodes.Unimplemented, ErrNotImplemented().Error())
+}